+++ app/containers/Root.dev.js
@@ -1,6 +1,5 @@
 import React, { Component, PropTypes } from 'react';
 import { Provider } from 'react-redux';
-import DevTools from './DevTools';
 import { Router } from 'react-router';
 import routes from '../routes';
 
@@ -12,7 +11,6 @@ export default class Root extends Component {
             <Provider store={store}>
                 <div>
                     <Router history={history} routes={routes} />
-                    <DevTools />
                 </div>
             </Provider>
         );
