+++ app/routes.js
@@ -2,8 +2,8 @@ import React from 'react';
 import { Route, IndexRoute } from 'react-router';
 import App from './components/App';
 import Home from './components/Home';
-import ProfileContainer from './containers/ProfileContainer';
-import ProjectRequest from './components/ProjectRequest';
+import ProfileContainer from './components/profile/ProfileContainer';
+import ProjectRequest from './components/projectrequest/ProjectRequest';
 
 
 export default (
